tools/image-generator/requirements.txt
anthropic>=0.40.0
openai>=1.0.0
replicate>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
